royalty -> splits. revert royalty.
